Decompose each permission view into it's only component
